chore(story assignments): minor refactor
